requirements.txt
numpy==2.2.2
pandas==2.2.3
torch==2.6.0
matplotlib==3.10.0
tqdm==4.67.1
scikit-learn==1.6.1
umap-learn==0.5.7
seaborn==0.13.1
pyperclip==1.9.0
statsmodels==0.14.4